fix: PR review 'updated' action, release download URLs, image loading, AuthUser Parcel bug
Ported/found while auditing a cluster of old crash fixes from fork
yxj-github-437 against our current code (most were already superseded -
these four weren't):

- ActivitiesAdapter/EventPayload: GitHub's PullRequestReviewEvent can carry
  action "updated" (also independently requested in upstream PR #380), which
  our enum didn't have - EventPayload.PullRequestReviewEventActionType.
  valueOf("updated") threw IllegalArgumentException immediately, crashing
  the activity feed whenever one appeared. Added the missing enum case.
- Release/DownloadSourceDialog: release source-code downloads now rewrite
  the API's tarball_url/zipball_url (api.github.com/repos/.../tarball/{ref})
  to the same direct codeload URL shape GitHub's own "Download ZIP" web
  button uses (github.com/{owner}/{repo}/archive/refs/tags/{tag}.{ext}),
  more reliable for a DownloadManager-based download than the redirect-based
  API URL.
- CodeWebView: loadImage()/setHtmlSource() used loadData() with no base URL,
  which gives WebView content an opaque origin and can silently fail to
  load remote resources (e.g. the image itself). Switched both to
  loadDataWithBaseURL via a shared loadHtml() helper, matching how
  setMdSource()/setCodeSource() already load content.
- AuthUser: writeToParcel wrote expireIn as a Long but the Parcel
  constructor read it back with readInt() - a real read/write mismatch that
  corrupts every field parcelled after it (scope, selected, loginId, name,
  avatar) whenever this object round-trips through a Bundle/Parcel (it's an
  @AutoAccess field on AppData, restored across process death). Found
  independently while checking a since-superseded fork commit titled "Fixed
  an issue with re-login" against our code. Fixed the constructor/setter to
  long and the Parcel read to readLong() to match.

Also confirmed already-present (no change needed): long-press a search
history entry to delete just that one already exists and works correctly
(SearchActivity's SearchRecordAdapter).

+++ app/src/main/java/com/thirtydegreesray/openhub/dao/AuthUser.java
@@ -41,7 +41,7 @@ public class AuthUser implements Parcelable {
     }
 
     @Generated
-    public AuthUser(String accessToken, java.util.Date authTime, int expireIn, String scope, boolean selected, String loginId, String name, String avatar) {
+    public AuthUser(String accessToken, java.util.Date authTime, long expireIn, String scope, boolean selected, String loginId, String name, String avatar) {
         this.accessToken = accessToken;
         this.authTime = authTime;
         this.expireIn = expireIn;
@@ -76,7 +76,7 @@ public class AuthUser implements Parcelable {
         return expireIn;
     }
 
-    public void setExpireIn(int expireIn) {
+    public void setExpireIn(long expireIn) {
         this.expireIn = expireIn;
     }
 
@@ -145,7 +145,7 @@ public class AuthUser implements Parcelable {
         this.accessToken = in.readString();
         long tmpAuthTime = in.readLong();
         this.authTime = tmpAuthTime == -1 ? null : new Date(tmpAuthTime);
-        this.expireIn = in.readInt();
+        this.expireIn = in.readLong();
         this.scope = in.readString();
         this.selected = in.readByte() != 0;
         this.loginId = in.readString();
